push to discord

+++ subalert/subq.py
@@ -1,5 +1,6 @@
-from .subtweet import Tweet
+from .discord import DiscordWebhook, DiscordAPI
 from .config import Configuration
+from .subtweet import Tweet
 import os
 
 config = Configuration()
@@ -45,9 +46,14 @@ class Queue:
 
         if 'batch_all' in self.items[0] and len(self.items[0]['batch_all']) >= 1 and self.items[0]['batch_all'][0] is not None:
             for tweet, media, collection_id in self.items[0]['batch_all']:
-                if not tweet:
-                    pass
 
+                # iterate over tweets that are blank.
+                # usually they're blank due to not matching conditions in set in config.local.yaml
+                if not tweet:
+                    continue
+
+                discord_webhook = DiscordWebhook(url=config.yaml_file['twitter']['sub_twitter']['NonFungibleTxs']['discord_webhook'])
+                discord_webhook.embeds(description=tweet, thumbnail='https://i.imgur.com/TO0jawi.png', footer='RMRK')
                 Tweet("NonFungibleTxs").alert(message=tweet, filename=media, verbose=True)
 
                 # Handle monitored collections
@@ -57,15 +63,19 @@ class Queue:
                 if collection_id:
                     monitored_collection = config.yaml_file['twitter']['collections'][collection_id]
                     account = list(monitored_collection.keys())[0]
+                    discord_webhook = DiscordWebhook(url=monitored_collection[account]['discord_webhook'])
+                    discord_webhook.embeds(description=tweet, thumbnail='https://i.imgur.com/TO0jawi.png', footer='RMRK')
                     Tweet(account, nft_collection=collection_id).alert(message=tweet, filename=media, verbose=True)
 
                 # only remove media if it actually returns anything.
-                if media:
+                if media and media != False:
                     os.remove(path=media)
 
         if 'transactions' in self.items[0] and len(self.items[0]['transactions']) >= 1:
             for tx in self.items[0]['transactions']:
                 if not tx:
-                    pass
+                    continue
 
+                discord_webhook = DiscordWebhook(url=config.yaml_file['twitter']['sub_twitter']['KusamaTxs']['discord_webhook'])
+                discord_webhook.embeds(description=tx, thumbnail='https://i.imgur.com/nkCOPOS.png', footer='Kusama Transaction')
                 Tweet("KusamaTxs").alert(message=tx, verbose=True)
